+++ internal/cfginit/cfginit.go
@@ -58,6 +58,17 @@ protoc_version: {{.ProtocVersion}}
 # Setting this will ignore unused imports.
 {{.V}}allow_unused_imports: true
 
+# Create directives.
+{{.V}}create:
+  # Map from relative directory to base package.
+  # This affects how packages are generated with create.
+  {{.V}}dir_to_base_package:
+    # This means that a file created "foo.proto" in the current directory will have package "bar".
+    # A file created "a/b/foo.proto" will have package "bar.a.b".
+    {{.V}}.:bar
+    # This means that a file created "idl/code.uber/a/b/c.proto" will have package "uber.a.b".
+    {{.V}}idl/code.uber:uber
+
 # Lint directives.
 {{.V}}lint:
   # Linter * files to ignore.